RAG-Powered Chatbot/requirements.txt
streamlit
python-dotenv
langchain
faiss-cpu
pypdf
langchain-community
tiktoken
openai